# api/openapi.yaml
openapi: "3.1.0"
info:
  title: DevSecOps Security Scan API
  version: "2.0.0"
  description: |
    ## Overview
    Full end-to-end REST API for the Jenkins-backed DevSecOps security scanning
    platform.  The API covers four functional domains:

    | Domain | Prefix | Description |
    |--------|--------|-------------|
    | **Scan** | `/scan` | Submit source code, track scan jobs, fetch results |
    | **Agent** | `/agent` | Provision / destroy dynamic Jenkins JNLP agents |
    | **Pipeline** | `/pipeline` | Trigger / query Jenkins pipeline executions |
    | **Reports** | `/reports` | Download generated security artifacts & SBOM |

    ### How it fits together
    ```
    Developer  ──POST /scan/upload──▶  API
                                        │
                                        ├─POST /agent/create────▶  Jenkins (dynamic JNLP)
                                        ├─POST /pipeline/trigger─▶  Jenkins Job
                                        ├─GET  /scan/{id}/status ─▶  poll build
                                        └─GET  /reports/{id}/...──▶  download artifacts
    ```
  contact:
    name: DevSecOps Platform Team
  license:
    name: MIT
servers:
  - url: http://132.186.17.22:9091
    description: Agent-host scan server (production)
  - url: http://localhost:9091
    description: Local development server
...
paths:
...
  /scan/upload:
    post:
      tags: [scan]
      operationId: upload_source
      summary: Upload source code archive
      description: |
        Accepts a gzipped tar archive of the project to scan.  
        The server stores it under `/opt/scan-uploads/<scan_id>/source.tar.gz`.  
        Maximum upload size: **1 GB**.
      parameters:
        - name: X-Scan-ID
          in: header
          required: true
          description: Unique scan identifier (e.g. `alice-laptop-1744000000`)
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_\-]{3,80}$'
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
              description: gzip-compressed tar archive of source tree
      responses:
        "200":
          description: Upload accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
              example:
                status: ok
                scan_id: alice-laptop-1744000000
                upload_path: /opt/scan-uploads/alice-laptop-1744000000
                size: 2097152
        "400":
          $ref: "#/components/responses/BadRequest"
        "413":
          description: Payload too large (> 1 GB)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Scan status & results ──────────────────────────────────
  /scan/{scan_id}/status:
    get:
      tags: [scan]
      operationId: get_scan_status
      summary: Poll scan / build status
      description: |
        Returns the current Jenkins build status for the given scan ID.
        Clients should poll every 5–10 seconds until `result` is one of
        `SUCCESS`, `FAILURE`, or `ABORTED`.
      parameters:
        - $ref: "#/components/parameters/ScanId"
      responses:
        "200":
          description: Current build status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanStatusResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /scan/{scan_id}/cancel:
    post:
      tags: [scan]
      operationId: cancel_scan
      summary: Abort a running scan
      description: Sends a stop signal to the Jenkins build associated with the scan.
      parameters:
        - $ref: "#/components/parameters/ScanId"
      responses:
        "200":
          description: Cancellation requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleOkResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /scan/{scan_id}/logs:
    get:
      tags: [scan]
      operationId: stream_scan_logs
      summary: Stream console log
      description: |
        Returns the full Jenkins console output captured so far.  
        Set `?start=<offset>` to paginate / tail the log.
      parameters:
        - $ref: "#/components/parameters/ScanId"
        - name: start
          in: query
          required: false
          description: Byte offset to resume from (for incremental polling)
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Console log text
          content:
            text/plain:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Cleanup ────────────────────────────────────────────────
  /scan/cleanup:
    post:
      tags: [scan]
      operationId: cleanup_scan
      summary: Remove uploaded source files
      description: |
        Deletes the uploaded tar archive and any temporary workspace files for
        the given scan ID. Called automatically by the client after the scan
        finishes, but can also be triggered manually.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanIdBody"
      responses:
        "200":
          description: Files removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleOkResponse"
              example:
                status: cleaned
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Agent lifecycle ────────────────────────────────────────
...
  /reports/{scan_id}:
    get:
      tags: [reports]
      operationId: list_report_files
      summary: List available report artifacts for a scan
      parameters:
        - $ref: "#/components/parameters/ScanId"
      responses:
        "200":
          description: List of artifact file paths
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportListResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /reports/{scan_id}/summary:
    get:
      tags: [reports]
      operationId: get_scan_summary
      summary: Get structured vulnerability summary
      description: |
        Returns a parsed JSON summary of all scan findings including:
        - Critical / High / Medium / Low counts by scan type
        - SBOM component count
        - Quality gate result
        - Secrets detected
      parameters:
        - $ref: "#/components/parameters/ScanId"
      responses:
        "200":
          description: Vulnerability summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanSummary"
        "404":
          $ref: "#/components/responses/NotFound"
  /reports/{scan_id}/download:
    get:
      tags: [reports]
      operationId: download_reports_archive
      summary: Download all reports as a tar.gz archive
      parameters:
        - $ref: "#/components/parameters/ScanId"
      responses:
        "200":
          description: Reports archive
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
  /reports/{scan_id}/{artifact}:
    get:
      tags: [reports]
      operationId: get_report_artifact
      summary: Download a specific report artifact
      description: |
        Fetch a specific file from the scan results, e.g.:
        - `secret-scan.json`
        - `trivy-fs-scan.json`
        - `sbom/trivy-cyclonedx-full.json`
        - `sbom/trivy-spdx-sbom.json`
        - `shellcheck.json`
        - `sonarqube-analysis.txt`
      parameters:
        - $ref: "#/components/parameters/ScanId"
        - name: artifact
          in: path
          required: true
          schema:
            type: string
          examples:
            secret:
              value: secret-scan.json
            trivy_fs:
              value: trivy-fs-scan.json
            sbom_cyclonedx:
              value: sbom/trivy-cyclonedx-full.json
      responses:
        "200":
          description: Artifact content
          content:
            application/json:
              schema: {}
            text/plain:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/NotFound"

# ─────────────────────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────────────────────
components:
...
  parameters:
    ScanId:
      name: scan_id
      in: path
      required: true
      description: Unique scan identifier (e.g. `alice-laptop-1744000000`)
      schema:
        type: string
        pattern: '^[a-zA-Z0-9_\-]{3,80}$'

  # ── Responses ───────────────────────────────────────────────
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server or downstream error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ── Schemas ─────────────────────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
        detail:
          type: string
          description: Optional additional context
    SimpleOkResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok, cleaned, cancelled]
    ScanIdBody:
      type: object
      required: [scan_id]
      properties:
        scan_id:
          type: string
          pattern: '^[a-zA-Z0-9_\-]{3,80}$'
...
    UploadResponse:
      type: object
      required: [status, scan_id, upload_path, size]
      properties:
        status:
          type: string
          enum: [ok]
        scan_id:
          type: string
        upload_path:
          type: string
        size:
          type: integer
          description: File size in bytes
    ScanStatusResponse:
      type: object
      required: [scan_id, building, result]
      properties:
        scan_id:
          type: string
        building:
          type: boolean
          description: True while the Jenkins build is still running
        result:
          type: string
          nullable: true
          enum: [SUCCESS, FAILURE, ABORTED, UNSTABLE, null]
          description: null while building
        build_number:
          type: integer
          nullable: true
        build_url:
          type: string
          format: uri
          nullable: true
        duration_ms:
          type: integer
          nullable: true
        estimated_duration_ms:
          type: integer
          nullable: true
        timestamp:
          type: integer
          description: Unix epoch millis of build start
    ReportListResponse:
      type: object
      required: [scan_id, artifacts]
      properties:
        scan_id:
          type: string
        artifacts:
          type: array
          items:
            type: string
          example:
            - secret-scan.json
            - trivy-fs-scan.json
            - trivy-sca.json
            - sbom/trivy-cyclonedx-full.json
            - sbom/trivy-spdx-sbom.json
            - shellcheck.json
            - sonarqube-analysis.txt
            - sonarqube-quality-gate.json
    ScanSummary:
      type: object
      required: [scan_id, gate_passed, totals]
      properties:
        scan_id:
          type: string
        gate_passed:
          type: boolean
        totals:
          $ref: "#/components/schemas/FindingTotals"
        findings_by_type:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/FindingTotals"
        sbom_component_count:
          type: integer
          nullable: true
        secrets_found:
          type: integer
          nullable: true
        sonarqube_quality_gate:
          type: string
          nullable: true
          enum: [OK, WARN, ERROR, null]
        pipeline:
          type: string
          nullable: true
        build_number:
          type: integer
          nullable: true
    FindingTotals:
      type: object
      properties:
        critical:
          type: integer
          default: 0
        high:
          type: integer
          default: 0
        medium:
          type: integer
          default: 0
        low:
          type: integer
          default: 0
        unknown:
          type: integer
          default: 0

    # Agent ────────────────────────────────────────────────────
...